Use the mruby-onig-regexp gem
We need a full-featured regex gem, and the mruby-pure-regexp gem we were
using previously (pulled in by our custom fork of mruby-marshal) turns
out to be missing some features.

We now explicitly include the Ongimo-based mruby-onig-regexp gem, which
has a full set of regex features. Doing so also allows us to use the
stock mruby-marshal gem, so this switches back to that.