Modify build to enable building mruby-onig-regexp and add arm64e support
The mruby-onig-regexp package depends on the Onigmo package, which is
built with autotools. In order to enable the entire package to build
completely, we need to use CrossBuild names that give the correct values
to autotools; by using standard 'triplet' names, autotools will be
configured correctly.

Another change is adding 'arm64e' as a new target--this lets us support
the newest iOS devices.

Finally, we bump the -miphoneos-version-min flag to 8.0, because it
wasn't building otherwise.